edit tags

+++ ImagePocket/ImagePocket/ImageCache.swift
@@ -143,12 +143,22 @@ final class ImageCache{
     }
     
     public func removeTagFromImages(tags: [TagEntity]) -> Void {
+        if tags.isEmpty {
+            return
+        }
         let removedTagImages = _imageRepository.remove(tags: tags)
         _tagCache.remove(tags: tags)
         let updatedImages = _imageRepository.getBy(ids: removedTagImages.map{$0.imageId})
         for image in updatedImages {
             _taggedImages[image.localIdentifier]?.removeTagId(entities: removedTagImages)
         }
+    }
+    
+    public func updateTagFromImages(tags: [TagEntity]) -> Void {
+        if tags.isEmpty {
+            return
+        }
+        _tagCache.saveOrUpdate(tags: tags)
     }
     
     private func addAssets(assets: [PHAsset]) -> Void {
